renamed registerRoute to userRoute
